recommendation_server/requirements.txt
fastapi
uvicorn
pymongo
pandas
scikit-learn
python-dotenv
google-generativeai
